+++ backend/src/database/db.py
@@ -127,7 +127,7 @@ class FarmDB:
         root_depth_cm: float,
         p_allowable: float,
         tree_count: int | None = None,
-        tree_height: int | None = None,
+        tree_height: float | None = None,
         row_distance: float | None = None,
         tree_distance: float | None = None,
         running_metre: float | None = None,
@@ -146,7 +146,7 @@ class FarmDB:
         area_ha_value = float(area_ha)
         p_allowable_value = float(p_allowable)
         tree_count_value = None if tree_count is None else int(tree_count)
-        tree_height_value = None if tree_height is None else int(tree_height)
+        tree_height_value = None if tree_height is None else float(tree_height)
         row_distance_value = None if row_distance is None else float(row_distance)
         tree_distance_value = None if tree_distance is None else float(tree_distance)
         running_metre_value = None if running_metre is None else float(running_metre)
